requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
groq
python-dotenv
PyPDF2
scikit-learn
joblib
nltk
matplotlib
seaborn